moved speed sections
