+++ components/SoftwareControl/speedSections.py
